third_party/blink/web_tests/external/wpt/html/semantics/the-button-element/command-and-commandfor/event-interface-expected.txt

This is a testharness.js-based test.
[FAIL] CommandEventInit properties set value
  assert_equals: expected Element node <body><div id="div"></div>\n<button id="button"></button>\n... but got null
[FAIL] CommandEventInit properties set value 2
  assert_equals: expected Element node <div id="div"></div> but got null
[FAIL] CommandEventInit properties set value 3
  assert_equals: expected Element node <button id="button"></button> but got null
Harness: the test ran to completion.
